Higuchi

+++ src_knn/inference.py
@@ -1,7 +1,7 @@
-import sklearn
+import pandas as pd
 
 
 def predict(model, json_input):
-    prediction = model.predict(json_input)
+    prediction = model.predict(json_input.astype(float))
     print(f"prediction : {prediction}")
     return prediction
